+++ data/src/main/java/com/example/data/db/local/db/SavedDatabase.kt
@@ -1,7 +1,9 @@
-package com.campusmap.android.wanted_preonboarding_android.roomdb
+package com.example.data.db.local.db
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
+import com.example.data.db.local.entity.Saved
+import com.example.data.db.local.dao.SavedDao
 
 // 필요하면 typeConverter까지
 @Database(entities = [Saved::class], version = 4)
