+++ data/src/main/java/com/example/data/db/local/entity/Saved.kt
@@ -1,4 +1,4 @@
-package com.campusmap.android.wanted_preonboarding_android.data.db.local.entity
+package com.example.data.db.local.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class Saved (
     @PrimaryKey(autoGenerate = true)
-    var id: Int? = null, // 요렇게하고 null로 주면 autoGenerate
+    var id: Int? = null,
     var author: String?,
     var content: String?,
     var publishedAt: String?,
