+++ data/src/main/java/com/example/data/db/remote/api/TopNewsApi.kt
@@ -1,6 +1,6 @@
-package com.campusmap.android.wanted_preonboarding_android.data.db.remote.api
+package com.example.data.db.remote.api
 
-import com.campusmap.android.wanted_preonboarding_android.data.db.remote.response.TopNewsResponse
+import com.example.data.db.remote.response.TopNewsResponse
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
@@ -13,7 +13,6 @@ interface TopNewsApi {
     @GET("top-headlines?country=us")
     suspend fun getTopNewsCategoryData(@Query("apiKey") apiKey : String,
                                        @Query("category") category: String) : Response<TopNewsResponse>
-
 
 
 }
